requirements.txt
pandas
python-dotenv
SQLAlchemy
psycopg2-binary
scikit-learn